requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# PostgreSQL (Neon)
# psycopg2-binary==2.9.9
psycopg2

# JWT Authentication
python-jose==3.3.0
python-dotenv==1.0.1

# Hashing
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Templates
jinja2==3.1.3
python-multipart==0.0.9

# Testing
pytest==8.1.1
httpx==0.27.0
